fix make error log

+++ files/initrd/opt/rr/include/modules.sh
@@ -42,7 +42,7 @@ function installModules() {
   local MLIST="${@}"
 
   if [ -z "${PLATFORM}" -o -z "${KVER}" ]; then
-    echo ""
+    echo "ERROR: installModules: Platform or Kernel Version not defined" >"${LOG_FILE}"
     return 1
   fi
   # Unzip modules for temporary folder
@@ -50,28 +50,37 @@ function installModules() {
   mkdir -p "${TMP_PATH}/modules"
   local KERNEL="$(readConfigKey "kernel" "${USER_CONFIG_FILE}")"
   if [ "${KERNEL}" = "custom" ]; then
-    tar -zxf "${CKS_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+    tar -zxf "${CKS_PATH}/modules-${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" 2>"${LOG_FILE}"
   else
-    tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules"
+    tar -zxf "${MODULES_PATH}/${PLATFORM}-${KVER}.tgz" -C "${TMP_PATH}/modules" 2>"${LOG_FILE}"
   fi
+  if [ $? -ne 0 ]; then
+    return 1
+  fi
+
   local ODP="$(readConfigKey "odp" "${USER_CONFIG_FILE}")"
   for F in $(ls "${TMP_PATH}/modules/"*.ko 2>/dev/null); do
     local M=$(basename ${F})
     [ "${ODP}" = "true" -a -f "${RAMDISK_PATH}/usr/lib/modules/${M}" ] && continue
     if echo "${MLIST}" | grep -wq "${M:0:-3}"; then
-      cp -f "${F}" "${RAMDISK_PATH}/usr/lib/modules/${M}"
+      cp -f "${F}" "${RAMDISK_PATH}/usr/lib/modules/${M}" 2>"${LOG_FILE}"
     else
-      rm -f "${RAMDISK_PATH}/usr/lib/modules/${M}"
+      rm -f "${RAMDISK_PATH}/usr/lib/modules/${M}" 2>"${LOG_FILE}"
     fi
   done
   mkdir -p "${RAMDISK_PATH}/usr/lib/firmware"
   if [ "${KERNEL}" = "custom" ]; then
-    tar -zxf "${CKS_PATH}/firmware.tgz" -C "${RAMDISK_PATH}/usr/lib/firmware"
+    tar -zxf "${CKS_PATH}/firmware.tgz" -C "${RAMDISK_PATH}/usr/lib/firmware" 2>"${LOG_FILE}"
   else
-    tar -zxf "${MODULES_PATH}/firmware.tgz" -C "${RAMDISK_PATH}/usr/lib/firmware"
+    tar -zxf "${MODULES_PATH}/firmware.tgz" -C "${RAMDISK_PATH}/usr/lib/firmware" 2>"${LOG_FILE}"
   fi
+  if [ $? -ne 0 ]; then
+    return 1
+  fi
+
   # Clean
   rm -rf "${TMP_PATH}/modules"
+  return 0
 }
 
 ###############################################################################
